Finish moving factory methods over to sub modules
